requirements.txt
Django==2.2.7
django-taggit==1.1.0
django-crispy-forms==1.8.1
Pillow==6.2.1
django-taggit==1.1.0